+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `--no-validate` flag to skip scope validation for `ado pat`. This allows addition of non-public ADO scopes.
 
 ## [0.8.4] - 2023-09-05
 ### Changed